Fixed buttons & added some styling

+++ public/js/fav.js
@@ -1,14 +1,16 @@
-const unsaveBtn = document.querySelectorAll('.unsaveBtn')
+const unsaveBtn = document.querySelector(".unsaveBtn");
 
-unsaveBtn.forEach((btn) => {
-    btn.addEventListener('click', (event) => {
-        let postId = event.currentTarget.parentElement.id
-        fetch(`/api/favorite/${postId}`, {
-            method: 'DELETE'
-        })
-        .then((response) => response.json())
-        .then((data) => console.log(data))
-        .then((window.location.reload()))
+unsaveBtn.addEventListener("click", (event) => {
+  if (event.target.classList.contains("unsaveBtn")) {
+    const postId = event.target.closest(".item-card").id;
+
+    fetch(`/api/favorite/${postId}`, {
+      method: "DELETE",
     })
-})
-
+      .then((response) => response.json())
+      .then((data) => {
+        console.log(data);
+        window.location.reload();
+      });
+  }
+});
